+++ CHANGES.md
@@ -3,6 +3,8 @@ Release Notes
 
 
 ### Other Changes
+- Allow `cls` parameter of `iter_all` to take an iterable of classes rather
+  than a single class.
 - Speed up `Part.iter_all()` when called without `cls`.  Although the order in
   which simultaneously starting/ending objects are returned is an implementation
   detail, note that this update may change that order with respect to previous